completing spanish translations
based on cleanup branch.

I choose valla over cerca for fences as in discussion with
natives it became clear that a cerca is an enclosing fence
which can also be called a valla. But a fence that does
not enclose entirely is not refered to as a cerca.
Since the player chooses what to do with the node, valla
covers both situations.

I also went with plural for stairs as there are multiple
steps in the single node.

Furthermore, changed all trunk names to Tronco from Madera
as this seems more precise.

With exception of the cocoa tree, all the items are
structured the same way now.

I left the no longer used entries but cleared out all other
comments as they are overkill for these mods and there were
many misplaced textdomain entries.